refactor(deeptrace): modularize pattern detectors
